Cut down on some unnecessarily long sleeps wherever possible.  This cuts the total unit test execution time down by about 50%.

|          | Debug    | Release  |
| -------- | -------- | -------- |
| Original | 4936 ms  | 3248 ms  |
| Current  | 2439 ms  | 1699 ms  |