@codemercenary 2015-10-27
Build server times are even more substantially improved.

| -------- | -------- |
| Original | 1560 ms  |
| Current  | 240 ms  |